Retain full original url in new location
Use the full original url in building the new location for redirection.

The previous version strips the mount point from the url when reconstructing it for redirection. 

For example, a call to http://servername/users?a=1&b=2, where /users is a router/app mount point, will get redirected to https://servername:port/?a=1&b=2 instead of https://servername:port/users?a=1&b=2.

This change fixes the problem.